feat: deep merge

+++ src/index.ts
@@ -29,10 +29,9 @@ export function persist(options: PersistOptions = {}) {
     const key = `${prefix}_${store.$id}`
 
     const storageResult = getItem(key, storage as Storage)
-    if (!storageResult || overwrite)
-      setItem(key, store.$state, storage)
+    if (!storageResult || overwrite) setItem(key, store.$state, storage)
     else
-      store.$patch(storageResult)
+      store.$patch(deepMerge(store.$state, storageResult))
 
     store.$subscribe(() => {
       setItem(key, store.$state, storage)
@@ -55,4 +54,29 @@ function getItem(key: string, storage: Storage): Record<string, any> | null {
 
 function setItem(key: string, state: unknown, storage: Storage) {
   return storage.setItem(key, JSON.stringify(state))
+}
+
+const isObject = (val: unknown) => val && typeof val === 'object'
+const mergeArrayWithDedupe = (a: any[], b: any[]) => Array.from(new Set([...a, ...b]))
+
+function deepMerge(
+  oldObj: Record<string | number | symbol, any>,
+  newObj: Record<string, any>,
+) {
+  const target: any = {}
+  for (const key of Object.keys(newObj)) {
+    const oldVal = oldObj[key]
+    const newVal = newObj[key]
+
+    if (Array.isArray(oldVal) && Array.isArray(newVal))
+      target[key] = mergeArrayWithDedupe(oldVal, newVal)
+
+    else if (isObject(oldVal) && isObject(newVal))
+      target[key] = deepMerge(oldVal, newVal)
+
+    else
+      target[key] = newVal
+  }
+
+  return { ...oldObj, ...target }
 }
